Implement step interface
An agent needs to traverse the environment, to this end the gym
interface exposes the step function. Tests have been added to make sure
it works with absolute and transformed values. Those were necessary,
particularly because the implementation needs to keep track of previous
and next observations. The reason why the next observation is cached, is
so that in case time series data is provided that can only be easily
traversed in one direction it still works.