Implement a log returns wrapper
Instead of configuring it via the constructor parameters, it is better
to wrap the environment via an ObservationWrapper. Not only does it
follow OpenAI gym conventions, it also makes it easier to combine it
with other observation wrappers. For instance one could combine it in
the future with a "FuzzObservationWrapper"